[core] -Fix: type typo in text geo swap caused truncation of x;y text scale factors
git-svn-id: svn://repo.hu/pcb-rnd/trunk@31903 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/obj_text.c
@@ -138,8 +138,8 @@ static int undo_text_geo_swap(void *udata)
 	pcb_poly_restore_to_poly(layer->parent.data, PCB_OBJ_TEXT, layer, g->text);
 
 	rnd_swap(int, g->Scale, g->text->Scale);
-	rnd_swap(int, g->scale_x, g->text->scale_x);
-	rnd_swap(int, g->scale_y, g->text->scale_y);
+	rnd_swap(double, g->scale_x, g->text->scale_x);
+	rnd_swap(double, g->scale_y, g->text->scale_y);
 	rnd_swap(rnd_coord_t, g->X, g->text->X);
 	rnd_swap(rnd_coord_t, g->Y, g->text->Y);
 	rnd_swap(rnd_coord_t, g->thickness, g->text->thickness);
